pyproject.toml
[tool.poetry]
name = "fastapi-zero"
version = "0.1.0"
description = ""
authors = ["Anderson Filho <filhoando@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.12,<4.0"
fastapi = {extras = ["standard"], version = "^0.123.0"}
sqlalchemy = "^2.0.45"
pydantic-settings = "^2.12.0"
alembic = "^1.17.2"
pwdlib = {extras = ["argon2"], version = "^0.3.0"}
tzdata = "^2025.3"
pyjwt = "^2.10.1"

[tool.poetry.group.dev.dependencies]
ruff = "^0.14.7"
pytest = "^9.0.1"
pytest-cov = "^7.0.0"
taskipy = "^1.14.1"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"


[dependency-groups]
dev = [
    "ruff (>=0.14.7,<0.15.0)",
    "pytest (>=9.0.1,<10.0.0)",
    "pytest-cov (>=7.0.0,<8.0.0)",
    "taskipy (>=1.14.1,<2.0.0)"
]

[tool.ruff]
line-length = 120
preview = true
extend-exclude = ["migrations"]

[tool.ruff.lint]
select = ["I", "F", "E", "FAST"]

[tool.pytest.ini_options]
pythonpath = "."
addopts = "-p no:warnings"
filterwarnings = [
    "ignore::ResourceWarning",
]

[tool.coverage.run]
# Isso força o coverage a olhar para a pasta correta, independente de como foi importado
source = ["fastapi_zero"]
omit = ["tests/*"]  # Garante que não vamos medir a cobertura dos próprios testes


[tool.taskipy.tasks]
lint = 'ruff check'
pre_format = 'ruff check --fix'
format = 'ruff format'
run = 'fastapi dev fastapi_zero/app.py'
pre_test = 'task lint'
test = 'pytest -s -x --cov=fastapi_zero -vv'
